@@ -1,15 +1,15 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.2"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.3"
 
     defaultConfig {
         applicationId "ru.ionov.timetable"
-        minSdkVersion 9
-        targetSdkVersion 23
-        versionCode 11
-        versionName "1.1.5"
+        minSdkVersion 16
+        targetSdkVersion 29
+        versionCode 12
+        versionName "2.0.0"
     }
     buildTypes {
         release {
@@ -17,9 +17,9 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
 
-        applicationVariants.all { variant ->
-            appendVersionName(variant, defaultConfig)
-        }
+//        applicationVariants.all { variant ->
+//            appendVersionName(variant, defaultConfig)
+//        }
     }
     packagingOptions {
         exclude 'META-INF/LICENSE'
@@ -27,32 +27,32 @@ android {
     }
 }
 
-def appendVersionName(variant, defaultConfig) {
-    variant.outputs.each { output ->
-        if (output.zipAlign) {
-            def file = output.outputFile
-            def fileName = "timetable-v" + defaultConfig.versionName  + ".apk";
-            output.outputFile = new File(file.parent, fileName)
-        }
-
-        def file = output.packageApplication.outputFile
-        def fileName = "timetable-release-v" + defaultConfig.versionName  + ".apk";
-        output.packageApplication.outputFile = new File(file.parent, fileName)
-    }
-}
+//def appendVersionName(variant, defaultConfig) {
+//    variant.outputs.each { output ->
+//        if (output.zipAlign) {
+//            def file = output.outputFile
+//            def fileName = "timetable-v" + defaultConfig.versionName  + ".apk";
+//            output.outputFile = new File(file.parent, fileName)
+//        }
+//
+//        def file = output.packageApplication.outputFile
+//        def fileName = "timetable-release-v" + defaultConfig.versionName  + ".apk";
+//        output.packageApplication.outputFile = new File(file.parent, fileName)
+//    }
+//}
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile 'com.android.support:cardview-v7:23.1.1'
-    compile 'com.android.support:recyclerview-v7:23.1.1'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    implementation 'androidx.appcompat:appcompat:1.1.0'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
+    implementation 'androidx.swiperefreshlayout:swiperefreshlayout:1.1.0-rc01'
 
-    compile 'com.github.rey5137:material:1.2.2'
+    implementation 'com.github.rey5137:material:1.2.5'
 
-    compile 'org.jsoup:jsoup:1.8.3'
-    compile 'com.nineoldandroids:library:2.4.0'
+    implementation 'org.jsoup:jsoup:1.8.3'
 
-    compile (
+    implementation (
             [group: 'com.fasterxml.jackson.core', name: 'jackson-core', version: '2.3.2'],
             [group: 'com.fasterxml.jackson.core', name: 'jackson-annotations', version: '2.3.2'],
             [group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.3.2']
