@@ -1,9 +1,8 @@
 package ru.ionov.timetable.activities;
 
 import android.os.Bundle;
-import android.support.v4.app.NavUtils;
-import android.support.v7.app.AppCompatActivity;
-
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.app.NavUtils;
 import ru.ionov.timetable.R;
 
 public class DateActivity extends AppCompatActivity
