@@ -1,18 +1,17 @@
 package ru.ionov.timetable.adapters;
 
-import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-import java.util.ArrayList;
-import java.util.List;
-
+import androidx.recyclerview.widget.RecyclerView;
 import ru.ionov.timetable.R;
 import ru.ionov.timetable.activities.CriteriaActivity;
 import ru.ionov.timetable.models.Criterion;
 import ru.ionov.timetable.viewholders.CriterionViewHolder;
+
+import java.util.ArrayList;
+import java.util.List;
 
 public class CriterionTileAdapter extends RecyclerView.Adapter<CriterionViewHolder>
 {
