@@ -2,18 +2,17 @@ package ru.ionov.timetable.adapters;
 
 import android.content.Context;
 import android.content.res.Resources;
-import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-
-import java.util.List;
-
+import androidx.recyclerview.widget.RecyclerView;
 import ru.ionov.timetable.R;
 import ru.ionov.timetable.models.Lesson;
 import ru.ionov.timetable.providers.PreferencesProvider;
 import ru.ionov.timetable.viewholders.LessonViewHolder;
+
+import java.util.List;
 
 public class LessonTileAdapter extends RecyclerView.Adapter<LessonViewHolder>
 {
