@@ -2,12 +2,10 @@ package ru.ionov.timetable.views;
 
 import android.content.Context;
 import android.graphics.Rect;
-import android.support.v4.view.ViewCompat;
-import android.support.v7.appcompat.BuildConfig;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.View;
+import androidx.core.view.ViewCompat;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 
 import java.lang.reflect.Field;
 
@@ -173,10 +171,10 @@ public class InnerRecyclerLayoutManager extends LinearLayoutManager
     }
 
     private void logMeasureWarning(int child) {
-        if (BuildConfig.DEBUG) {
-            Log.w("LinearLayoutManager", "Can't measure child #" + child + ", previously used dimensions will be reused." +
-                    "To remove this message either use #setChildSize() method or don't run RecyclerView animations");
-        }
+//        if (BuildConfig.DEBUG) {
+//            Log.w("LinearLayoutManager", "Can't measure child #" + child + ", previously used dimensions will be reused." +
+//                    "To remove this message either use #setChildSize() method or don't run RecyclerView animations");
+//        }
     }
 
     private void initChildDimensions(int width, int height, boolean vertical) {
@@ -270,8 +268,8 @@ public class InnerRecyclerLayoutManager extends LinearLayoutManager
 
     private static void onMakeInsertDirtyFailed() {
         canMakeInsetsDirty = false;
-        if (BuildConfig.DEBUG) {
-            Log.w("LinearLayoutManager", "Can't make LayoutParams insets dirty, decorations measurements might be incorrect");
-        }
+//        if (BuildConfig.DEBUG) {
+//            Log.w("LinearLayoutManager", "Can't make LayoutParams insets dirty, decorations measurements might be incorrect");
+//        }
     }
 }
