@@ -1,8 +1,8 @@
 package ru.ionov.timetable.views;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class SwipeRefreshRecyclerView extends RecyclerView
 {
